ZERO-27 [Rename]: common 폴더 @shared 로 변경
